+++ plugins/modules/zhmc_nic.py
@@ -187,7 +187,7 @@ EXAMPLES = """
 # Note: The following examples assume that some variables named 'my_*' are set.
 
 - name: Ensure NIC exists in the partition
-  zhmc_partition:
+  zhmc_nic:
     hmc_host: "{{ my_hmc_host }}"
     hmc_auth: "{{ my_hmc_auth }}"
     cpc_name: "{{ my_cpc_name }}"
@@ -202,7 +202,7 @@ EXAMPLES = """
   register: nic1
 
 - name: Ensure NIC does not exist in the partition
-  zhmc_partition:
+  zhmc_nic:
     hmc_host: "{{ my_hmc_host }}"
     hmc_auth: "{{ my_hmc_auth }}"
     cpc_name: "{{ my_cpc_name }}"
@@ -211,7 +211,7 @@ EXAMPLES = """
     state: absent
 
 - name: Gather facts about a NIC
-  zhmc_partition:
+  zhmc_nic:
     hmc_host: "{{ my_hmc_host }}"
     hmc_auth: "{{ my_hmc_auth }}"
     cpc_name: "{{ my_cpc_name }}"